Reviving DrainCredit(), adding back into Receiver (#356)
Adding the drain function back.

One of the big questions from before was what was and was not required by the spec. The behavior of flow+drain is a "SHOULD" behavior, not a "MUST", which gives implementations wiggle room. Rather than exclude the function, we've decided to include it but make sure that the caveats are well-documented.

For reference, this is the AMQP text for the drain flag:

> The drain flag indicates how the sender SHOULD behave when insufficient messages are available to consume the current link-credit. If set, the sender will (after sending all available messages) advance the delivery-count as much as possible, consuming all link-credit and send the flow state to the receiver. Only the receiver can independently modify this field. The sender's value is always the last known value indicated by the receiver.